apps/api/pyproject.toml
[project]
name = "task-garden-api"
version = "0.1.0"
description = "FastAPI backend scaffold for Task Garden."
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.6",
  "httpx>=0.28.1",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.7.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.1"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]

[tool.alembic]
script_location = "app/db/migrations"
